% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{add_compound}
\alias{add_compound}
\title{Add one or more compounds to a snapshot}
\usage{
add_compound(snapshot, compound)
}
\arguments{
\item{snapshot}{A \link{Snapshot} object.}

\item{compound}{A \link{Compound} object created with \code{\link[=create_compound]{create_compound()}}, or a
list of such objects.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Add one or more \link{Compound} objects to a \link{Snapshot}.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  add_compound(create_compound(name = "Drug X"))

snapshot <- load_snapshot("Midazolam") |>
  add_compound(list(
    create_compound(name = "Drug X"),
    create_compound(name = "Drug Y")
  ))
}
}
