% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{add_formulation}
\alias{add_formulation}
\title{Add one or more formulations to a snapshot}
\usage{
add_formulation(snapshot, formulation)
}
\arguments{
\item{snapshot}{A Snapshot object}

\item{formulation}{A Formulation object created with
\code{\link[=create_formulation]{create_formulation()}}, or a list of such objects.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Add one or more \link{Formulation} objects to a \link{Snapshot}.
}
\examples{
\dontrun{
# Load a snapshot
snapshot <- load_snapshot("Midazolam")

# Add a single formulation
form <- create_formulation(name = "Tablet", type = "Weibull")
snapshot <- add_formulation(snapshot, form)

# Add several at once
forms <- list(
  create_formulation("Tablet", type = "Weibull"),
  create_formulation("Oral solution", type = "First Order")
)
snapshot <- add_formulation(snapshot, forms)
}
}
