% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{AdvancedParameter}
\alias{AdvancedParameter}
\title{AdvancedParameter class for Population advanced parameters}
\description{
An R6 class that represents an advanced parameter in a population.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{The raw data of the parameter (read-only)}

\item{\code{name}}{The name of the parameter}

\item{\code{seed}}{The seed used for parameter generation}

\item{\code{distribution_type}}{The distribution type of the parameter}

\item{\code{parameters}}{The parameters of the distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdvancedParameter-new}{\code{AdvancedParameter$new()}}
\item \href{#method-AdvancedParameter-print}{\code{AdvancedParameter$print()}}
\item \href{#method-AdvancedParameter-clone}{\code{AdvancedParameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedParameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedParameter-new}{}}}
\subsection{Method \code{new()}}{
Create a new AdvancedParameter object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedParameter$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw parameter data from a population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AdvancedParameter object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedParameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedParameter-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the advanced parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the AdvancedParameter object for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedParameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedParameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedParameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
