% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ospsuite_time_to_duration}
\alias{convert_ospsuite_time_to_duration}
\title{Convert time value and unit to lubridate duration}
\usage{
convert_ospsuite_time_to_duration(value, unit)
}
\arguments{
\item{value}{The time value}

\item{unit}{The ospsuite time unit}
}
\value{
A lubridate duration object
}
\description{
Convert a time value and ospsuite unit to a lubridate duration object.
This function handles unit conversion and special cases like kiloseconds.
}
