% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ospsuite_time_unit_to_lubridate}
\alias{convert_ospsuite_time_unit_to_lubridate}
\title{Convert ospsuite time units to lubridate-compatible units}
\usage{
convert_ospsuite_time_unit_to_lubridate(unit)
}
\arguments{
\item{unit}{The ospsuite time unit to convert}
}
\value{
A lubridate-compatible time unit
}
\description{
Convert time units from ospsuite format to lubridate-compatible format.
This function handles the mapping of units like "day(s)" to "days", etc.
}
