% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_schema.R
\name{create_output_schema}
\alias{create_output_schema}
\title{Create an output schema for a simulation}
\usage{
create_output_schema(intervals = list())
}
\arguments{
\item{intervals}{List of \link{OutputInterval} objects (created with
\code{\link[=create_output_interval]{create_output_interval()}}) or raw lists. Defaults to an empty list.}
}
\value{
An \link{OutputSchema} object.
}
\description{
Build an \link{OutputSchema} from a list of \link{OutputInterval} objects.
}
\examples{
create_output_schema(
  intervals = list(
    create_output_interval(start_time = 0, end_time = 2, resolution = 20),
    create_output_interval(start_time = 2, end_time = 24, resolution = 20)
  )
)
}
