% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{Event}
\alias{Event}
\title{Event class for OSP snapshot events}
\description{
An R6 class that represents an event in an OSP snapshot.
This class provides methods to access different properties of an event
and display a summary of its information.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{The raw data of the event (read-only)}

\item{\code{name}}{The name of the event}

\item{\code{template}}{The template of the event}

\item{\code{parameters}}{The list of parameter objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Event-new}{\code{Event$new()}}
\item \href{#method-Event-print}{\code{Event$print()}}
\item \href{#method-Event-to_dataframe}{\code{Event$to_dataframe()}}
\item \href{#method-Event-clone}{\code{Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-new"></a>}}
\if{latex}{\out{\hypertarget{method-Event-new}{}}}
\subsection{Method \code{new()}}{
Create a new Event object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw event data from a snapshot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Event object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-print"></a>}}
\if{latex}{\out{\hypertarget{method-Event-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-Event-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Convert the event to a data frame for easier manipulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble with event information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Event-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
