% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_dataframes.R
\name{get_expressions_df}
\alias{get_expressions_df}
\title{Get expression profile data for all individuals in a snapshot}
\usage{
get_expressions_df(snapshot)
}
\arguments{
\item{snapshot}{A snapshot object}
}
\value{
A tibble containing expression profile data for all individuals
}
\description{
Extract expression profile data from all individuals in a snapshot and combine them into a single data frame.
}
