% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_dataframes.R
\name{get_origin_df}
\alias{get_origin_df}
\title{Get origin data for all individuals in a snapshot}
\usage{
get_origin_df(snapshot)
}
\arguments{
\item{snapshot}{A snapshot object}
}
\value{
A tibble containing origin data for all individuals
}
\description{
Extract origin data from all individuals in a snapshot and combine them into a single data frame.
}
