% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_dataframes.R
\name{get_parameters_df}
\alias{get_parameters_df}
\title{Get parameter data for all individuals in a snapshot}
\usage{
get_parameters_df(snapshot)
}
\arguments{
\item{snapshot}{A snapshot object}
}
\value{
A tibble containing parameter data for all individuals
}
\description{
Extract parameter data from all individuals in a snapshot and combine them into a single data frame.
}
