% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formulation.R
\name{load_formulations}
\alias{load_formulations}
\title{Load formulations from a list}
\usage{
load_formulations(formulation_list)
}
\arguments{
\item{formulation_list}{List. List of formulation data from a snapshot}
}
\value{
List of Formulation objects
}
\description{
Converts a list of formulation data to Formulation objects
}
\examples{
\dontrun{
# Load snapshot and get formulations
snapshot <- load_snapshot("path/to/snapshot.json")
formulations <- load_formulations(snapshot$data$Formulations)
}
}
