% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{load_populations}
\alias{load_populations}
\title{Load populations from snapshot data}
\usage{
load_populations(population_list)
}
\arguments{
\item{population_list}{List of population data from a snapshot}
}
\value{
A named list of Population objects
}
\description{
Load populations from a list of population data from a snapshot
}
