% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite-package.R
\docType{package}
\name{osp.snapshots-package}
\alias{osp.snapshots}
\alias{osp.snapshots-package}
\title{osp.snapshots: Import and Manipulate Open Systems Pharmacology Project Snapshots}
\description{
Provides a convenient R interface for working with PKSIM project snapshots exported as JSON files. You provide the snapshot file, and 'osp.snapshots' handles the complex nested structure, allowing you to easily import, navigate, modify, and export snapshot data. Building blocks can be converted to data frames for integration with the 'ospsuite' R framework and broader R ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://esqlabs.github.io/osp.snapshots/}
}

}
\author{
\strong{Maintainer}: Felix MIL \email{felix.mil@esqlabs.com}

}
\keyword{internal}
