% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterCollection.R
\name{ParameterCollection}
\alias{ParameterCollection}
\title{ParameterCollection class for managing a collection of parameters}
\description{
An R6 class that manages a collection of Parameter objects.
Provides methods to get, set, and add parameters.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{Named list of Parameter objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParameterCollection-new}{\code{ParameterCollection$new()}}
\item \href{#method-ParameterCollection-get_parameter}{\code{ParameterCollection$get_parameter()}}
\item \href{#method-ParameterCollection-set_parameter}{\code{ParameterCollection$set_parameter()}}
\item \href{#method-ParameterCollection-add_parameter}{\code{ParameterCollection$add_parameter()}}
\item \href{#method-ParameterCollection-print}{\code{ParameterCollection$print()}}
\item \href{#method-ParameterCollection-clone}{\code{ParameterCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-new}{}}}
\subsection{Method \code{new()}}{
Create a new ParameterCollection object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$new(parameters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{List of raw parameter data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ParameterCollection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get a parameter by path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$get_parameter(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character. The path of the parameter to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Parameter object or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set a parameter by path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$set_parameter(path, parameter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character. The path of the parameter to set}

\item{\code{parameter}}{Parameter object to set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-add_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-add_parameter}{}}}
\subsection{Method \code{add_parameter()}}{
Add a new parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$add_parameter(parameter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{Parameter object to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
