% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.biliary_clearance}
\alias{print.biliary_clearance}
\title{Print method for biliary clearance processes}
\usage{
\method{print}{biliary_clearance}(x, ...)
}
\arguments{
\item{x}{A biliary_clearance object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for biliary clearance processes
}
