% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.compound_collection}
\alias{print.compound_collection}
\title{S3 print method for compound collections}
\usage{
\method{print}{compound_collection}(x, ...)
}
\arguments{
\item{x}{A compound collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the compound collection
}
\description{
S3 print method for compound collections
}
