% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.compound_processes}
\alias{print.compound_processes}
\title{S3 print method for compound processes}
\usage{
\method{print}{compound_processes}(x, ...)
}
\arguments{
\item{x}{A compound_processes object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the compound processes
}
\description{
S3 print method for compound processes
}
