% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.event_collection}
\alias{print.event_collection}
\title{S3 print method for event collections}
\usage{
\method{print}{event_collection}(x, ...)
}
\arguments{
\item{x}{An event collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the event collection
}
\description{
S3 print method for event collections
}
