% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.expression_profile_collection}
\alias{print.expression_profile_collection}
\title{S3 print method for expression profile collections}
\usage{
\method{print}{expression_profile_collection}(x, ...)
}
\arguments{
\item{x}{An expression profile collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the expression profile collection
}
\description{
S3 print method for expression profile collections
}
