% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.formulation_collection}
\alias{print.formulation_collection}
\title{S3 print method for formulation collections}
\usage{
\method{print}{formulation_collection}(x, ...)
}
\arguments{
\item{x}{A formulation collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the formulation collection
}
\description{
S3 print method for formulation collections
}
