% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.hepatic_clearance}
\alias{print.hepatic_clearance}
\title{Print method for hepatic clearance processes}
\usage{
\method{print}{hepatic_clearance}(x, ...)
}
\arguments{
\item{x}{A hepatic_clearance object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for hepatic clearance processes
}
