% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.individual_collection}
\alias{print.individual_collection}
\title{S3 print method for individual collections}
\usage{
\method{print}{individual_collection}(x, ...)
}
\arguments{
\item{x}{An individual collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the individual collection
}
\description{
S3 print method for individual collections
}
