% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.induction}
\alias{print.induction}
\title{Print method for induction processes}
\usage{
\method{print}{induction}(x, ...)
}
\arguments{
\item{x}{An induction object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for induction processes
}
