% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.inhibition}
\alias{print.inhibition}
\title{Print method for inhibition processes}
\usage{
\method{print}{inhibition}(x, ...)
}
\arguments{
\item{x}{An inhibition object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for inhibition processes
}
