% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.metabolizing_enzymes}
\alias{print.metabolizing_enzymes}
\title{Print method for metabolizing enzymes}
\usage{
\method{print}{metabolizing_enzymes}(x, ...)
}
\arguments{
\item{x}{A metabolizing_enzymes object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for metabolizing enzymes
}
