% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.observed_data_collection}
\alias{print.observed_data_collection}
\title{Print method for observed data collection}
\usage{
\method{print}{observed_data_collection}(x, n = 5, ...)
}
\arguments{
\item{x}{An observed_data_collection object (a named list of DataSet objects)}

\item{n}{Number of observed data names to display (default: 10)}

\item{...}{Additional arguments (not used)}
}
\value{
The input object x, invisibly
}
\description{
Print method for observed data collection
}
