% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.population_collection}
\alias{print.population_collection}
\title{S3 print method for population collections}
\usage{
\method{print}{population_collection}(x, ...)
}
\arguments{
\item{x}{A population collection object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the population collection
}
\description{
S3 print method for population collections
}
