% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.protein_binding_partners}
\alias{print.protein_binding_partners}
\title{Print method for protein binding partners}
\usage{
\method{print}{protein_binding_partners}(x, ...)
}
\arguments{
\item{x}{A protein_binding_partners object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for protein binding partners
}
