% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.renal_clearance}
\alias{print.renal_clearance}
\title{Print method for renal clearance processes}
\usage{
\method{print}{renal_clearance}(x, ...)
}
\arguments{
\item{x}{A renal_clearance object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for renal clearance processes
}
