% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.transporter_proteins}
\alias{print.transporter_proteins}
\title{Print method for transporter proteins}
\usage{
\method{print}{transporter_proteins}(x, ...)
}
\arguments{
\item{x}{A transporter_proteins object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for transporter proteins
}
