% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{remove_compound}
\alias{remove_compound}
\title{Remove compounds from a snapshot}
\usage{
remove_compound(snapshot, compound_name)
}
\arguments{
\item{snapshot}{A \link{Snapshot} object.}

\item{compound_name}{Character vector of compound names to remove.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Remove one or more compounds from a \link{Snapshot} by name.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  remove_compound("Midazolam")
}
}
