% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{remove_event}
\alias{remove_event}
\title{Remove events from a snapshot}
\usage{
remove_event(snapshot, event_name)
}
\arguments{
\item{snapshot}{A [Snapshot] object.}

\item{event_name}{Character vector of event names to remove.}
}
\value{
The updated [Snapshot] object, returned invisibly.
}
\description{
Remove one or more events from a [Snapshot] by name.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  remove_event("Breakfast")
}
}
