% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{remove_protocol}
\alias{remove_protocol}
\title{Remove protocols from a snapshot}
\usage{
remove_protocol(snapshot, protocol_name)
}
\arguments{
\item{snapshot}{A [Snapshot] object.}

\item{protocol_name}{Character vector of protocol names to remove.}
}
\value{
The updated [Snapshot] object, returned invisibly.
}
\description{
Remove one or more protocols from a [Snapshot] by name.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  remove_protocol("IV bolus 1mg")
}
}
