% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{remove_simulation}
\alias{remove_simulation}
\title{Remove simulations from a snapshot}
\usage{
remove_simulation(snapshot, simulation_name)
}
\arguments{
\item{snapshot}{A \link{Snapshot} object.}

\item{simulation_name}{Character vector of simulation names to remove.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Remove one or more simulations from a \link{Snapshot} by name.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  remove_simulation("simulation1")
}
}
