% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_gender}
\alias{validate_gender}
\title{Validate that a gender is valid}
\usage{
validate_gender(gender)
}
\arguments{
\item{gender}{Gender value to validate}
}
\value{
TRUE if valid, throws an error if invalid
}
\description{
Check if a gender is valid using ospsuite Gender enum.
Throws an error if the gender is invalid.
}
