% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_population}
\alias{validate_population}
\title{Validate that a population is valid}
\usage{
validate_population(population)
}
\arguments{
\item{population}{Population value to validate}
}
\value{
TRUE if valid, throws an error if invalid
}
\description{
Check if a population is valid using ospsuite HumanPopulation enum.
Throws an error if the population is invalid.
}
