% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_snapshot}
\alias{validate_snapshot}
\title{Validate that an object is a Snapshot}
\usage{
validate_snapshot(snapshot)
}
\arguments{
\item{snapshot}{Object to validate}
}
\value{
TRUE if valid, throws an error if invalid
}
\description{
Check if an object is a Snapshot instance.
Throws an error if the object is not a Snapshot.
}
