% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_species}
\alias{validate_species}
\title{Validate that a species is valid}
\usage{
validate_species(species)
}
\arguments{
\item{species}{Species value to validate}
}
\value{
TRUE if valid, throws an error if invalid
}
\description{
Check if a species is valid using ospsuite Species enum.
Throws an error if the species is invalid.
}
