% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_unit}
\alias{validate_unit}
\title{Validate that a unit is valid for a given dimension}
\usage{
validate_unit(unit, dimension)
}
\arguments{
\item{unit}{Unit to validate}

\item{dimension}{Dimension the unit should belong to}
}
\value{
TRUE if valid, throws an error if invalid
}
\description{
Check if a unit is valid for a given dimension using ospsuite units.
Throws an error if the unit is invalid.
}
