% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{add_compound}
\alias{add_compound}
\title{Add a compound to a snapshot}
\usage{
add_compound(snapshot, compound)
}
\arguments{
\item{snapshot}{A [Snapshot] object.}

\item{compound}{A [Compound] object created with [create_compound()].}
}
\value{
The updated [Snapshot] object, returned invisibly.
}
\description{
Add a [Compound] object to a [Snapshot].
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  add_compound(create_compound(name = "Drug X"))
}
}
