% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{add_event}
\alias{add_event}
\title{Add an event to a snapshot}
\usage{
add_event(snapshot, event)
}
\arguments{
\item{snapshot}{A \link{Snapshot} object.}

\item{event}{An \link{Event} object created with \code{\link[=create_event]{create_event()}}.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Add an \link{Event} object to a \link{Snapshot}.
}
\examples{
\dontrun{
evt <- create_event(
  name = "Breakfast",
  template = "Meal: Standard (Human)"
)
snapshot <- load_snapshot("Midazolam") |>
  add_event(evt)
}
}
