% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{add_protocol}
\alias{add_protocol}
\title{Add a protocol to a snapshot}
\usage{
add_protocol(snapshot, protocol)
}
\arguments{
\item{snapshot}{A \link{Snapshot} object.}

\item{protocol}{A \link{Protocol} object created with \code{\link[=create_protocol]{create_protocol()}}.}
}
\value{
The updated \link{Snapshot} object, returned invisibly.
}
\description{
Add a \link{Protocol} object to a \link{Snapshot}.
}
\examples{
\dontrun{
prot <- create_protocol(
  name = "Single oral dose",
  application_type = "Oral",
  dosing_interval = "Single"
)
snapshot <- load_snapshot("Midazolam") |>
  add_protocol(prot)
}
}
