% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionProfile.R
\name{ExpressionProfile}
\alias{ExpressionProfile}
\title{ExpressionProfile class for OSP snapshot expression profiles}
\description{
An R6 class that represents an expression profile in an OSP snapshot.
This class provides methods to access different properties of an expression profile
and display a summary of its information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The raw data of the expression profile}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{The type of the expression profile}

\item{\code{species}}{The species of the expression profile}

\item{\code{molecule}}{The molecule name of the expression profile}

\item{\code{category}}{The category of the expression profile}

\item{\code{localization}}{The localization of the expression profile}

\item{\code{ontogeny}}{The ontogeny information of the expression profile}

\item{\code{parameters}}{The parameters of the expression profile}

\item{\code{id}}{The unique identifier of the expression profile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExpressionProfile-new}{\code{ExpressionProfile$new()}}
\item \href{#method-ExpressionProfile-print}{\code{ExpressionProfile$print()}}
\item \href{#method-ExpressionProfile-to_df}{\code{ExpressionProfile$to_df()}}
\item \href{#method-ExpressionProfile-clone}{\code{ExpressionProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExpressionProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExpressionProfile-new}{}}}
\subsection{Method \code{new()}}{
Create a new ExpressionProfile object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpressionProfile$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw expression profile data from a snapshot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ExpressionProfile object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExpressionProfile-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExpressionProfile-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the expression profile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpressionProfile$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExpressionProfile-to_df"></a>}}
\if{latex}{\out{\hypertarget{method-ExpressionProfile-to_df}{}}}
\subsection{Method \code{to_df()}}{
Convert the expression profile to data frames for easier manipulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpressionProfile$to_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with basic information and parameters as tibbles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExpressionProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExpressionProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpressionProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
