% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{remove_observed_data}
\alias{remove_observed_data}
\title{Remove observed data from a snapshot}
\usage{
remove_observed_data(snapshot, observed_data_name)
}
\arguments{
\item{snapshot}{A [Snapshot] object.}

\item{observed_data_name}{Character vector of observed-data names to
remove.}
}
\value{
The updated [Snapshot] object, returned invisibly.
}
\description{
Remove one or more observed-data entries from a [Snapshot] by name.
}
\examples{
\dontrun{
snapshot <- load_snapshot("Midazolam") |>
  remove_observed_data("Study A")
}
}
