% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{Snapshot}
\alias{Snapshot}
\title{Snapshot class for OSP snapshots}
\description{
An R6 class that represents an OSP snapshot file. This class provides
methods to access different components of the snapshot and visualize its structure.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The raw data of the snapshot}

\item{\code{path}}{The path to the snapshot file}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pksim_version}}{The human-readable PKSIM version corresponding to the snapshot version}

\item{\code{compounds}}{List of Compound objects in the snapshot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Snapshot-new}{\code{Snapshot$new()}}
\item \href{#method-Snapshot-print}{\code{Snapshot$print()}}
\item \href{#method-Snapshot-export}{\code{Snapshot$export()}}
\item \href{#method-Snapshot-clone}{\code{Snapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-new}{}}}
\subsection{Method \code{new()}}{
Create a new Snapshot object from a JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the snapshot JSON file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Snapshot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the snapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-export"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-export}{}}}
\subsection{Method \code{export()}}{
Export the snapshot to a JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to save the JSON file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
