% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snapshot.R
\name{Snapshot}
\alias{Snapshot}
\title{Snapshot class for OSP snapshots}
\description{
An R6 class that represents an OSP snapshot file. This class provides
methods to access different components of the snapshot and visualize its structure.
}
\examples{

## ------------------------------------------------
## Method `Snapshot$add_individual`
## ------------------------------------------------

\dontrun{
# Create a new individual
ind <- create_individual(name = "New Patient", age = 35, weight = 70)

# Add the individual to a snapshot
snapshot$add_individual(ind)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{The path to the snapshot file if loaded from file, NULL otherwise}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{The aggregated data of the snapshot from all components}

\item{\code{pksim_version}}{The human-readable PKSIM version corresponding to the snapshot version}

\item{\code{compounds}}{List of Compound objects in the snapshot}

\item{\code{individuals}}{List of Individual objects in the snapshot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Snapshot-new}{\code{Snapshot$new()}}
\item \href{#method-Snapshot-print}{\code{Snapshot$print()}}
\item \href{#method-Snapshot-export}{\code{Snapshot$export()}}
\item \href{#method-Snapshot-add_individual}{\code{Snapshot$add_individual()}}
\item \href{#method-Snapshot-clone}{\code{Snapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-new}{}}}
\subsection{Method \code{new()}}{
Create a new Snapshot object from a JSON file or a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$new(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Path to the snapshot JSON file, URL, template name, or a list containing snapshot data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Snapshot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the snapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-export"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-export}{}}}
\subsection{Method \code{export()}}{
Export the snapshot to a JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to save the JSON file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-add_individual"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-add_individual}{}}}
\subsection{Method \code{add_individual()}}{
Add an Individual object to the snapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$add_individual(individual)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{An Individual object created with create_individual()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Create a new individual
ind <- create_individual(name = "New Patient", age = 35, weight = 70)

# Add the individual to a snapshot
snapshot$add_individual(ind)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
